+++ src/QueryCat.Backend/Formatters/FormattersInfo.cs
@@ -1,6 +1,6 @@
 using QueryCat.Backend.Abstractions;
-using QueryCat.Backend.Storage;
-using QueryCat.Backend.Utils;
+using QueryCat.Backend.Execution;
+using QueryCat.Backend.Functions;
 
 namespace QueryCat.Backend.Formatters;
 
@@ -9,47 +9,60 @@ namespace QueryCat.Backend.Formatters;
 /// </summary>
 public static class FormattersInfo
 {
-    private static readonly Dictionary<string, Func<IRowsFormatter>> Formatters = new()
+    private static readonly Dictionary<string, Func<ExecutionThread, IRowsFormatter>> Formatters = new()
     {
         // File extensions.
-        [".csv"] = () => new DsvFormatter(),
-        [".tsv"] = () => new DsvFormatter('\t'),
-        [".tab"] = () => new DsvFormatter('\t'),
-        [".log"] = () => new DsvFormatter(new StreamRowsInputOptions
-        {
-            DelimiterStreamReaderOptions = new DelimiterStreamReader.ReaderOptions
-            {
-                Delimiters = new[] { ' ' },
-                DelimitersCanRepeat = true,
-                SkipEmptyLines = true,
-            }
-        }),
-        [".json"] = () => new JsonFormatter(),
-        [".xml"] = () => new XmlFormatter(),
-        [".xsd"] = () => new XmlFormatter(),
+        [".csv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
+        [".tsv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
+                FunctionArguments.Create().Add("delimiter", '\t')),
+        [".tab"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
+                FunctionArguments.Create().Add("delimiter", '\t')),
+        [".log"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
+                FunctionArguments.Create().Add("delimiter", ' ').Add("delimiter_can_repeat", false)),
+        [".json"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("json"),
+        [".xml"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
+        [".xsd"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
 
         // Content types.
-        ["text/csv"] = () => new DsvFormatter(),
-        ["text/x-csv"] = () => new DsvFormatter(','),
-        ["application/csv"] = () => new DsvFormatter(','),
-        ["application/x-csv"] = () => new DsvFormatter(','),
-        ["text/tab-separated-values"] = () => new DsvFormatter('\t'),
-        ["application/json"] = () => new JsonFormatter(),
-        ["application/xml"] = () => new XmlFormatter(),
-        ["application/xhtml+xml"] = () => new XmlFormatter(),
-        ["application/soap+xml"] = () => new XmlFormatter(),
+        ["text/csv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
+        ["text/x-csv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
+        ["application/csv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
+        ["application/x-csv"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv"),
+        ["text/tab-separated-values"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("csv",
+                FunctionArguments.Create().Add("delimiter", '\t')),
+        ["application/json"] = thread=>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("json"),
+        ["application/xml"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
+        ["application/xhtml+xml"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
+        ["application/soap+xml"] = thread =>
+            thread.FunctionsManager.CallFunction<IRowsFormatter>("xml"),
     };
 
     /// <summary>
     /// Create formatter by file extension or content type.
     /// </summary>
     /// <param name="id">File extension or content type.</param>
+    /// <param name="thread">Execution thread.</param>
     /// <returns>Instance of <see cref="IRowsFormatter" /> or null.</returns>
-    public static IRowsFormatter? CreateFormatter(string id)
+    public static IRowsFormatter? CreateFormatter(string id, ExecutionThread thread)
     {
         if (Formatters.TryGetValue(id.ToLower(), out var factory))
         {
-            return factory.Invoke();
+            return factory.Invoke(thread);
         }
         return null;
     }
@@ -59,7 +72,7 @@ public static class FormattersInfo
     /// </summary>
     /// <param name="id">Identifier (file extension or content type).</param>
     /// <param name="formatterFunc">Delegate to create formatter.</param>
-    public static void RegisterFormatter(string id, Func<IRowsFormatter> formatterFunc)
+    public static void RegisterFormatter(string id, Func<ExecutionThread, IRowsFormatter> formatterFunc)
     {
         Formatters[id] = formatterFunc;
     }