+++ src/QueryCat.Backend/Providers/FileInputOutput.cs
@@ -1,6 +1,7 @@
 using System.ComponentModel;
 using System.IO.Compression;
 using QueryCat.Backend.Abstractions;
+using QueryCat.Backend.Execution;
 using QueryCat.Backend.Formatters;
 using QueryCat.Backend.Functions;
 using QueryCat.Backend.Types;
@@ -19,8 +20,8 @@ internal static class FileInputOutput
     public static VariantValue ReadFile(FunctionCallInfo args)
     {
         var path = args.GetAt(0).AsString;
-        var formatter = args.Count > 1 ? args.GetAt(1).AsObject as IRowsFormatter : GetFormatter(path);
-        var files = GetFileInputsByPath(path, formatter).ToList();
+        var formatter = args.Count > 1 ? args.GetAt(1).AsObject as IRowsFormatter : GetFormatter(path, args.ExecutionThread);
+        var files = GetFileInputsByPath(path, args.ExecutionThread, formatter).ToList();
         if (!files.Any())
         {
             throw new QueryCatException($"No files match '{path}'.");
@@ -41,7 +42,7 @@ internal static class FileInputOutput
         }
 
         var formatter = args.GetAt(1).AsObject as IRowsFormatter;
-        formatter ??= GetFormatter(path);
+        formatter ??= GetFormatter(path, args.ExecutionThread);
         Stream file = File.Open(path, FileMode.Create, FileAccess.Write, FileShare.ReadWrite);
         if (CompressFilesExtensions.Contains(Path.GetExtension(path).ToLower()))
         {
@@ -50,11 +51,12 @@ internal static class FileInputOutput
         return VariantValue.CreateFromObject(formatter.OpenOutput(file));
     }
 
-    private static IEnumerable<IRowsInput> GetFileInputsByPath(string path, IRowsFormatter? formatter = null)
+    private static IEnumerable<IRowsInput> GetFileInputsByPath(string path,
+        ExecutionThread thread, IRowsFormatter? formatter = null)
     {
         foreach (var file in GetFilesByPath(path))
         {
-            var fileFormatter = formatter ?? GetFormatter(file);
+            var fileFormatter = formatter ?? GetFormatter(file, thread);
             Stream fileStream = File.Open(file, FileMode.Open, FileAccess.Read, FileShare.ReadWrite);
             if (CompressFilesExtensions.Contains(Path.GetExtension(file).ToLower()))
             {
@@ -94,14 +96,14 @@ internal static class FileInputOutput
         }
     }
 
-    private static IRowsFormatter GetFormatter(string path)
+    private static IRowsFormatter GetFormatter(string path, ExecutionThread thread)
     {
         var extension = Path.GetExtension(path).ToLower();
         if (CompressFilesExtensions.Contains(extension))
         {
             extension = Path.GetExtension(path.Substring(0, path.Length - extension.Length)).ToLower();
         }
-        var formatter = FormattersInfo.CreateFormatter(extension);
+        var formatter = FormattersInfo.CreateFormatter(extension, thread);
         return formatter ?? new TextLineFormatter();
     }
 }