+++ src/QueryCat.Backend/Types/VariantValue.cs
@@ -358,12 +358,50 @@ public readonly partial struct VariantValue : IEquatable<VariantValue>
     #endregion
 
     /// <summary>
-    /// Get object of specified type.
+    /// Get object of the specified type.
     /// </summary>
     /// <typeparam name="T">Type.</typeparam>
     /// <returns>Object.</returns>
-    public T GetAsObject<T>()
+    public T As<T>()
     {
+        var retType = typeof(T);
+
+        if (retType == typeof(Int64)
+            || retType == typeof(Int32)
+            || retType == typeof(Int16)
+            || retType == typeof(Byte)
+            || retType == typeof(UInt64)
+            || retType == typeof(UInt32)
+            || retType == typeof(UInt16)
+            || retType == typeof(SByte))
+        {
+            return (T)Convert.ChangeType(AsInteger, typeof(T));
+        }
+        if (retType == typeof(bool))
+        {
+            return (T)Convert.ChangeType(AsBoolean, typeof(T));
+        }
+        if (retType == typeof(double) || retType == typeof(float))
+        {
+            return (T)Convert.ChangeType(AsFloat, typeof(T));
+        }
+        if (retType == typeof(decimal))
+        {
+            return (T)Convert.ChangeType(AsNumeric, typeof(T));
+        }
+        if (retType == typeof(string))
+        {
+            return (T)Convert.ChangeType(AsString, typeof(T));
+        }
+        if (retType == typeof(DateTime) || retType == typeof(DateTimeOffset))
+        {
+            return (T)Convert.ChangeType(AsTimestamp, typeof(T));
+        }
+        if (retType == typeof(TimeSpan))
+        {
+            return (T)Convert.ChangeType(AsInterval, typeof(T));
+        }
+
         var sourceObj = CheckTypeAndTryToCast(DataType.Object)._object;
         if (sourceObj == null)
         {