+++ CHANGELOG.md
@@ -9,10 +9,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add "--no-header", "--float-format" command line arguments.
 - Add "NoSpaceTable" output style.
 - Support strings unescape.
+- Add more arguments to function "csv".
 
 ### Fixed
 
 - Fix file parsing without a new line at the end.
+
 
 ## [0.4.5] - 2023-05-29